feat: add auth context
