+++ src/common/common.context.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+import { AuthContext, AuthContextDefault } from './common.types';
+export const Context = createContext<AuthContext>(AuthContextDefault);
