+++ src/common/common.types.ts
@@ -49,3 +49,13 @@ export const boardDefault: BoardInfo = {
   description: '',
   columns: [],
 };
+
+export type AuthContext = {
+  isAuthorize: boolean;
+  setIsAuthorize: (value: boolean) => void;
+};
+
+export const AuthContextDefault: AuthContext = {
+  isAuthorize: false,
+  setIsAuthorize: () => {},
+};
