+++ src/components/Header/Header.scss
@@ -111,11 +111,11 @@
 }
 
 .language__btn:hover,
-.language__btn[data-active="true"] {
+.language__btn[data-active='true'] {
   color: #00deb6;
 }
 
-.language__btn[data-active="true"] {
+.language__btn[data-active='true'] {
   font-weight: 700;
 }
