+++ src/i18n/en/ns1.json
@@ -26,7 +26,7 @@
     "tit": "Title"
   },
   "edit_Profile": {
-    "title": "Edit Profile",
+    "title": "Profile",
     "edit_profile": "Edit Profile",
     "name": "Create Your Name",
     "login": "Create Your Login",
